requirements.txt
Flask==3.0.0
google-generativeai==0.8.3
mysql-connector-python==8.2.0
bcrypt==4.1.2
cryptography==45.0.5
